Add Potentiometer Control and Enhance UI/UX Features
This commit introduces several enhancements to the IoT control application, focusing on adding support for potentiometer control and improving the overall user interface and experience. The key changes include:

1. **Potentiometer Control Integration**:
   - Added a new `PotentiometerCubit` to manage the state of the potentiometer value fetched from Firebase.
   - Created a `PotentiometerBlocBuilder` widget to display the potentiometer value using a custom half-circle progress indicator.
   - Integrated the potentiometer control into the main IoT control view and updated the UI layout to accommodate the new feature.

2. **UI/UX Enhancements**:
   - Introduced a new `boxDecoration` function to apply a gradient background to the main control view, enhancing visual appeal.
   - Modified the `CustomHalfCircleProgress` widget to accept an integer percentage value, improving its usability for displaying potentiometer values.
   - Reorganized the layout of the IoT control view to better structure the display of various control elements, including LEDs, fan, and potentiometer.

3. **Firebase Data Handling**:
   - Updated the `fetchDataFromFirebase` function to handle potentiometer data, ensuring real-time updates are reflected in the application.
   - Streamlined the Firebase data fetching process to include the new potentiometer field, maintaining consistency across all control elements.

4. **Speech-to-Text Functionality**:
   - Enhanced the speech-to-text functionality to control IoT devices based on voice commands, providing a more intuitive user experience.
   - Updated the voice command handling to include specific commands for controlling the potentiometer, expanding the range of voice-controlled features.

These changes collectively improve the functionality and usability of the IoT control application, providing a more comprehensive and user-friendly interface for managing connected devices.

+++ lib/core/widgets/custom_half_circle_progress.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 
 class CustomHalfCircleProgress extends StatelessWidget {
-  final double percentage;
+  final int  percentage;
 
   CustomHalfCircleProgress({required this.percentage});
 
@@ -17,7 +17,7 @@ class CustomHalfCircleProgress extends StatelessWidget {
 }
 
 class HalfCircleProgressPainter extends CustomPainter {
-  final double percentage;
+  final int percentage;
 
   HalfCircleProgressPainter({required this.percentage});
